Bug fixed in remove_answer_images_by_q_id function. Previously if no answer images were found by question_id, logical error occured. Now even if no answers contained any images, program runs without exception.

+++ logic.py
@@ -186,11 +186,12 @@ def remove_answer_images_by_q_id(question_id):
     """
     images = fetch_answer_images(question_id)
 
-    for image in images:
-        try:
-            os.remove('static/uploads/' + image)
-        except (FileNotFoundError, TypeError):
-            pass
+    if images:
+        for image in images:
+            try:
+                os.remove('static/uploads/' + image)
+            except (FileNotFoundError, TypeError):
+                pass
 
 
 def create_new_answer_no_image(message, question_id):
